+++ bot/__init__.py
@@ -1,1 +1,10 @@
 """Пакет с обработчиками бота."""
+
+from . import add_handlers, update_handlers, search_handlers, utils
+
+__all__ = [
+    "add_handlers",
+    "update_handlers",
+    "search_handlers",
+    "utils",
+]
